Split du fichier sql_api/crud
