+++ src/sql_api/__init__.py
@@ -1,17 +1,11 @@
-from .creds import Creds
-from .crud import (
+from .allow import (
     allow_domain_for_user,
-    create_domain,
-    create_user,
-    delete_user,
     deny_domain_for_user,
     get_allowed,
     get_allows,
-    get_domain,
-    get_domains,
-    get_user,
-    get_users,
-    nb_users,
 )
+from .creds import Creds
 from .database import get_db, get_maker, init_db
+from .domain import create_domain, get_domain, get_domains
 from .models import DBAllowed, DBDomain, DBUser
+from .user import create_user, delete_user, get_user, get_users, nb_users
